requirements.txt
# Data science and ML essentials
numpy>=1.26.2
pandas>=2.1.4
matplotlib>=3.8.2
seaborn>=0.13.0
scikit-learn>=1.3.2

# Utility packages
tqdm>=4.66.1

# LLM interaction
llama-cpp-python>=0.2.90
bitsandbytes
evaluate
rouge-score
psutil
